+++ bemuse/src/previewer/PreviewFileDropHandler.tsx
@@ -1,0 +1,19 @@
+import { useEffect } from 'react'
+
+export const PreviewFileDropHandler = () => {
+  useEffect(() => {
+    const onDragOver = (e: DragEvent) => {
+      e.preventDefault()
+    }
+    const onDrop = (e: DragEvent) => {
+      e.preventDefault()
+    }
+    window.addEventListener('dragover', onDragOver)
+    window.addEventListener('drop', onDrop)
+    return () => {
+      window.removeEventListener('dragover', onDragOver)
+      window.removeEventListener('drop', onDrop)
+    }
+  }, [])
+  return null
+}
